+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -432,6 +432,7 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             ? null
             : FlexSchemeVariant.values[controller.usedFlexToneSetup]
                 .tones(Brightness.dark)
+                .monochromeSurfaces(controller.useMonoSurfacesDark)
                 .onMainsUseBW(controller.onMainsUseBWDark)
                 .onSurfacesUseBW(controller.onSurfacesUseBWDark)
                 .surfacesUseBW(controller.surfacesUseBWDark),